requirements.txt
pandas
scikit-learn
imblearn
shap
matplotlib
seaborn
fpdf2
joblib